% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabufuncs.R
\docType{data}
\name{tabuObj}
\alias{tabuObj}
\title{Base Tabu Search Function}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
tabuObj
}
\arguments{
\item{Takes}{in a tabuSearch object}

\item{Object}{is a tabuSearch object, Verbose is a boolean which if true returns a more detailed summary}

\item{Takes}{in a tabuSearch object}

\item{Object}{is a tabuSearch object, Verbose is a boolean which if true returns a more detailed summary}

\item{Takes}{in a tabuSearch object}

\item{Object}{is a tabuSearch object, Verbose is a boolean which if true returns a more detailed summary}

\item{Takes}{in a tabuSearch object}

\item{Object}{is a tabuSearch object, Verbose is a boolean which if true returns a more detailed summary}
}
\description{
Base Tabu Search Function
}
\examples{
tabuSearch(x)
Summary function for base tabuSearch

summ(x,T)
Proximity Tabu Search Function

tabuSearchProx(x)
Summary function for tabuSearchProx

ProxSumm(x,T)
Knapsack Tabu Search Function

tabuSearchKnap(x)
Summary function for tabuSearchKnap

KnapSumm(x,T)
Travelling Sales Person Tabu Search Function

tabuSearchTSP(x)
Summary function for tabuSearchTSP

TSPSumm(x,T)
}
\keyword{Knapsack}
\keyword{Proximity}
\keyword{Summary}
\keyword{TSP}
\keyword{tabuSearch}
